+++ src/main/java/mario/russo/infra/UsuarioPanache.java
@@ -5,7 +5,7 @@ import jakarta.enterprise.context.ApplicationScoped;
 import mario.russo.core.domain.entity.UsuarioEntity;
 
 @ApplicationScoped
-public class UsuarioPanache implements PanacheRepositoryBase<UsuarioEntity, Long> {
+public class UsuarioPanache implements PanacheRepositoryBase<UsuarioEntity, Integer> {
     public UsuarioEntity findByEmail(String email) {
         return find("email", email).firstResult();
     }
